openldap: Fix CVE-2017-14159
This commit add the description of why CVE-2017-14159 is being discarted
as invalid. The CVE-2017-14159 reports that slapd in OpenLDAP 2.4.45 and
earlier creates a PID file after dropping privileges to a non-root
account, which might allow local users to kill arbitrary processes by
leveraging access to this non-root account for PID file modification
before a root script executes a "kill `cat /pathname`" command, as
demonstrated by openldap-initscript.

Acording to http://www.openldap.org/its/index.cgi?findid=8703
there was a confusion the wat  that PIDs get recycled. The security
shoudl be on init scripts which is not related to  the OpenLDAP Project.

This is why only a cve-2017-14159.nopatch was added